bots/examples/requirements.txt
telethon==1.34.0
python-telegram-bot==20.8
discord.py==2.3.2
